When used in Production, JS wasn't loading
There seems to be a problem using this with a Rails app in
Production mode; everything is fine in development and
staging but in production the chartkick.js file was
served as an empty file but with a 200. This looks like a
result of a misconfig of the Sinatra assetpack, possibly,
and so this branch uses a simplified strategy which is
working happily in Production.

(Graphs are awesome)